requirements.txt
Django==3.1.4
gunicorn==20.0.4
psycopg2==2.8.6
redis==3.5.3
django-progressive-web-app==0.1.1
django-geojson==3.1.0
django-leaflet==0.27.1
django-pwa==1.0.10
jsonfield==3.1.0
pip==20.3.1
pytz==2020.4
setuptools==51.0.0
six==1.15.0
sqlparse==0.4.1
serviceworker~=1.0.4
djangorestframework~=3.12.2
djangorestframework-gis~=0.17
jedi~=0.18.0